% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lz4_compress}
\alias{lz4_compress}
\title{Compress a raw vector using LZ4}
\usage{
lz4_compress(input)
}
\arguments{
\item{input}{A \code{raw} vector to compress.}
}
\value{
A \code{raw} vector of LZ4-compressed bytes.
}
\description{
Compress a raw vector using LZ4
}
\keyword{internal}
