% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lz4_decompress}
\alias{lz4_decompress}
\title{Decompress an LZ4-compressed raw vector}
\usage{
lz4_decompress(input, uncompressed_size)
}
\arguments{
\item{input}{A \code{raw} vector of LZ4-compressed bytes.}

\item{uncompressed_size}{The expected size of the decompressed output in bytes.}
}
\value{
A \code{raw} vector of decompressed bytes.
}
\description{
Decompress an LZ4-compressed raw vector
}
\keyword{internal}
